HomePage Realtime Update Done

+++ app/src/main/java/com/teamdui/profiler/ui/History/progress.java
@@ -4,7 +4,7 @@ public class progress {
     public int cal;
     public int wat;
     public int exr;
-    public int calburn;
+    public double calburn;
 
     public progress() {
         cal = 0;
